Create shared structure family class for server and client
