+++ tiled/structures/core.py
@@ -1,0 +1,15 @@
+"""
+This module was created to save structure methods that can be handled by
+the server and the client.
+
+"""
+
+import enum
+
+
+class StructureFamilies(str, enum.Enum):
+    node = "node"
+    array = "array"
+    dataframe = "dataframe"
+    xarray_data_array = "xarray_data_array"
+    xarray_dataset = "xarray_dataset"
